requirements.txt
asgiref==3.5.2
attrs==22.1.0
beautifulsoup4==4.11.1
boto3==1.24.59
botocore==1.27.59
branca==0.5.0
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.5
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
dj-static==0.0.6
Django==4.1
django-filter==22.1
docopt==0.6.2
Fiona @ file:///C:/Users/w7147/pipwin/Fiona-1.8.21-cp310-cp310-win_amd64.whl
folium==0.12.1.post1
GDAL @ file:///C:/Users/w7147/pipwin/GDAL-3.4.3-cp310-cp310-win_amd64.whl
geopandas==0.11.1
gunicorn==20.1.0
idna==3.3
Jinja2==3.1.2
jmespath==1.0.1
Js2Py==0.71
Markdown==3.4.1
MarkupSafe==2.1.1
munch==2.5.0
numpy==1.23.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
pipwin==0.5.2
psycopg2==2.9.3
pycparser==2.21
pyimgur==0.6.0
pyjsparser==2.7.1
PyJWT==2.4.0
pyparsing==3.0.9
PyPrind==2.11.3
pyproj==3.4.0
pySmartDL==1.3.4
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
Shapely==1.8.4
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
static3==0.7.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.11
whitenoise==6.2.0
